Fix: search results table -- issue with handling array data input (#1260)
* Fix: search results table -- issue with handling array data input
name-examination-code-with-us phase-1

as a work-arround for this issue, changed  all instances of the table being fed an array of data with string-templates so the table no longer receives any data as an array.

* fix: add-cancel-comment-modal opening when making decision in error
namex-ui-enhancements code-with-us
per comments of pr 1260

usage of bootstrap3 data-* attributes in mixed bootstrap/vuetify layout resulted in data-* attributes being applied to incorrect components. Decision buttons were opening Cancel Modal.  added a method for opening and closing the modal using jQuery and removed data-* attributes from mark-up.